Added command line arg for directory

+++ wallpaper.py
@@ -1,3 +1,6 @@
+#!/usr/local/bin/python
+
+import argparse
 import os
 import wget
 import re
@@ -9,8 +12,14 @@ class ArsenalWallpaperExtractor(object):
     def __init__(self):
         self.file = ""
         self.arsenal_wallpapers_link = "http://www.arsenal.com/fanzone/wallpapers"
-        self.wallpaper_directory = "/Users/Adithya/Desktop/Stuff/Arsenal/2014-15" # TODO: should parameterize
+        self.wallpaper_directory = self._get_wallpaper_directory()
         self.latest_picture_number = self._get_latest_picture_number()
+
+    def _get_wallpaper_directory(self):
+        directory = os.path.abspath(args.dir)
+        if not os.path.isdir(directory):
+            os.makedirs(directory)
+        return directory
 
     def _get_latest_picture_number(self):
         files = os.listdir(self.wallpaper_directory)
@@ -20,17 +29,19 @@ class ArsenalWallpaperExtractor(object):
             m = re.search(r'gun__(\d+)_3.jpg', filename)
             if m:
                 picture_numbers.append(m.group(1))
+        if not picture_numbers:
+            return 0 # TODO: Still need to fix bugs for directory with no new pictures
         picture_numbers.sort(reverse=True)
         return picture_numbers[0]
 
-    def download_html_file(self):
+    def _download_html_file(self):
         # TODO: check if file already exists
         self.file = wget.download(self.arsenal_wallpapers_link, bar=None)
 
-    def delete_html_file(self):
+    def _delete_html_file(self):
         os.remove(os.path.join(os.getcwd(), self.file))
 
-    def parse_html_file(self):
+    def _parse_html_file(self):
         with open(self.file, 'r') as f:
             file_contents = f.read()
         regex = '/assets/_files/desktops/\w+_\d+/gun__(\d+)_3.jpg'
@@ -38,7 +49,7 @@ class ArsenalWallpaperExtractor(object):
         picture_numbers.sort(reverse=True)
         return picture_numbers
 
-    def get_newer_pictures(self, picture_numbers):
+    def _get_newer_pictures(self, picture_numbers):
         if picture_numbers[0] <= self.latest_picture_number:
             print('No new pictures found!')
             return
@@ -47,21 +58,25 @@ class ArsenalWallpaperExtractor(object):
             if picture_number > self.latest_picture_number:
                 picture_name = wget.download('http://www.arsenal.com/assets/_files/desktops/jul_14/gun__%s_3.jpg' % picture_number, bar=None)
                 logging.info('Copying picture %s to %s ....', picture_name, self.wallpaper_directory)
-                shutil.copy(picture_name, self.wallpaper_directory)
-                os.remove(os.path.join(os.getcwd(), picture_name))
+                if self._not_current_directory():
+                    shutil.copy(picture_name, self.wallpaper_directory)
+                    os.remove(os.path.join(os.getcwd(), picture_name))
                 picture_count += 1
             else:
                 break
         print('%d pictures downloaded!' % picture_count)
 
-    def extract_wallpapers(self):
-        self.download_html_file()
-        self.get_newer_pictures(self.parse_html_file())
-        self.delete_html_file()
+    def _not_current_directory(self):
+        return os.getcwd() != self.wallpaper_directory
+
+    def _extract_wallpapers(self):
+        self._download_html_file()
+        self._get_newer_pictures(self._parse_html_file())
+        self._delete_html_file()
 
 def get_new_wallpapers():
     extractor = ArsenalWallpaperExtractor()
-    extractor.extract_wallpapers()
+    extractor._extract_wallpapers()
 
 def main():
     print 'Extracting wallpapers from arsenal.com .....' 
@@ -69,6 +84,11 @@ def main():
 
 # set logging level and log file
 logging.basicConfig(format='%(asctime)s - %(levelname)s - %(message)s', filename='wallpaper.log', level=logging.DEBUG)
+
+# parse command line arguments
+parser = argparse.ArgumentParser(description='Tool to obtain new arsenal.com wallpapers')
+parser.add_argument('-d', '--dir',  required=True, help='Directory in which wallpapers will be downloaded to.')
+args = parser.parse_args()
 
 if __name__ == "__main__":
     main()
